+++ PROJECT.py
@@ -33,31 +33,14 @@ class PROJECT(PROJECT_BASE):
     FIXME: list[str] = [
     ]
     # -----------------------------------------------------------------------------------------------------------------
-    VERSION = Version((0, 2, 14))
+    VERSION = Version((0, 2, 15))
     NEWS: list[str] = [
-        "[ini] add ConfigParserMod",
-        "[AttrKit] add",
-        [
-            "[Text]",
-                "add parse__dict_ini",
-                "fix__json/parse__json",
-                "add delete__cmts_c",
-                "add search__group - need fix",
-         ],
-        "[TextFile] add DictTextFileLoader",
-        "[NestInit_AttrsLambdasResolve] use with Resolvers",
-        ["[Nest*_Attrs]",
-            "separate all for Attrs Contains/Eq/Len/StR",
-            "[NestSAI_AttrAnycase] deprecate all Setattrs* cause recursion exx",
-
-         ],
-        ["[EqValid]",
-            "add EqValid_AnnotsAllExists",
-            "add EqValid_AttrsByObjNotPrivate*",
-         ],
-        "[PV] deprecate old and add new PvLoaderIni/json/env",
-        "[Kwargs] ref to simple dict nesting",
-
+        "[TESTS] move all near source code",
+        ["[AnnotAttrAux] BIG REF",
+            "make nested Annots from Attrs +use same meth in both",
+            "rename sai_*/gai_",
+            "add values__reinit_mutable +NestInit_Mutable",
+        ],
     ]